requirements.txt
bcrypt
pyjwt
python-dotenv
requests
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pydantic
greenlet
cryptography